Notify Panopticon when a browse page is created
This changes the MainstreamBrowsePagesController so that, when a browse
page is created in this app, a request is made to Panopticon to create
the browse page there too.

The notifier is invoked as part of the create action, and will fail 
noisily if the request is unsuccessful, as the Panopticon API adapter
will raise an exception. For now, it makes sense not to catch this, as 
consistency between this app and Panopticon is the most important thing.