Tidy up list of mainstream browse pages
This makes the list of mainstream browse pages a table, and changes the
data displayed so that only parent tags are listed here.

Here, we're using RSpec's feature spec pattern to have integration tests
for smaller aspects of the app which fall outside of the 'critical path'
scope for the Cucumber test suite.