Allow child browse pages to be created
This changes the existing mainstream browse pages implementation so that
child browse pages can be created under a parent browse page.

This also changes the 'show' view for parent browse pages, so that they 
display a list of their children, and add a link to the form to create
a new child browse page.

A child browse page is created through the same 'new'/'create' actions
as parent browse pages, but with a parent_id attribute, provided through
a URL parameter (and checked during the rendering of the page).

Panopticon is notified of the creation of child browse pages in the same
way as parents.