Register prefix routes for curated content.
This change was already made in Panopticon to
support the /latest and /email-signup pages, but
not updated here.  Since Panopticon doesn't use
the URL arbiter for this, that means that
Collections Publisher can sometimes overwrite the
prefix route with an exact route, taking the sub-
pages offline (see https://govuk.zendesk.com/agent/tickets/936651).

The intention is for these to change to multiple
exact routes when topic creation is moved from
Panopticon into Collections Publisher.