Don't allow changing Tag slugs once published
Once a `Tag` has been published, do not allow changing the slug.

The implementation for this is largely covered by
`ActiveModel::Dirty`[1], which allows tracking changes to attributes on
an `ActiveRecord` model by exposing methods such as
`attribute_changed?`. These methods can then be used as part of a
validation.

[1] http://api.rubyonrails.org/classes/ActiveModel/Dirty.html