Make User.uid NOT NULL.
We should never have a user without a uid, so it makes sense to set this
ad NOT NULL.  Doing this exposed an issue in the tests where our user
factory didn't set a uid.  This meant that the warden session
serialisation wasn't working leading to failures when we introduced some
javascript integration tests.