Allow associating topics with mainstream browse
The only thing to be aware of is that the controller `update` method
here tries to map over the submitted `Array` of ids (which are `String`
values) and tries to find the respective `Topic` objects.

This is because of how the association between `Topic` and
`MainstreamBrowsePage` models work using the `TagAssociation`. Rails
isn't clever enough with it's magic to deal with this cleanly.