Upgrade aasm to 4.1
To keep up to date.  The change from version 3 to 4 has changed several
things, mostly about the ordering of callbacks etc.  See
https://github.com/aasm/aasm/blob/master/README_FROM_VERSION_3_TO_4.md
for full details.

It also introduces a `no_direct_assignment` option to prevent direct
writes to the state column, which means that we can remove our manual
method of achieving the same thing.