Fix test failure github reporting.
This was reporting that the run errored instead of failed.  This is
because the error handler was being invoked any time a command exited
non-zero.

Changing this to run the tests as part of an if statement prevents the
error handler being triggered, and allows us to handle the outcome
apropriately.