Re-layout list editor
The aim of this PR is to improve the UX of the list editor.

- The editor now has a left hand side with all items tagged to the
item. On the right all lists are displayed. This hopefully improves the
understanding of the relationship between the two parts.
- Items from the left hand side are only "draggable". Previously it was
possible to sort and drag items in this list, causing confusion. It
also allows users to add a link to multiple lists, something which
could previously only be achieved by turning off javascript.
- When an item is included in one or more lists it is bold
- When the user is about to create the first list for a tag we show a
special page explaining they're about to change the "sort mode" of the
tag. This can be further improved by setting the ordering explicitly -
but that's for a later PR.
- The layout of the tables is updated to fit more items on the screen
and make dragging easier.
- Previously, when an item was dragged we calculated the exact rows of
that had to be updated. For example, when the user dragged the fifth
item in a list of 10 to the fourth place, we would update items 4
through 10. Finding out which items to update turned out to be a
fruitful source of bugs and confusion. This behaviour has been changed
to always update all items in a list on each update. In the previous
example, instead of 4-10, we update 1-10. In a later PR, we should use
bulk-updating to the server (which is now not very hard to implement).
- Related, keeping data in sync is hard on JS.
GOVUK.curatedLists.showEmptyLabels and
GOVUK.curatedLists.redrawUncategorizedList are purposely "naive". They
look at the state of the world and re-render it on each update, instead
of trying to only update things that changed, which would be more
performant.