Make TopicPresenter updated_at tests more explicit
save! doesn't actually update the updated_at field if there are no
changes to be saved.  These tests currently only pass because the topic
variable is lazily loaded.

This changes these tests to be explicit about the update time they
expect, and also to use `touch` which will always update the timestamp
on the record.