Don't default to calculated lists during publishing
We don't want to default to `TagPresenter#build_groups` when the tag
doesn't have a `published_groups` set yet.

The conditional was in there because in the time between the deploy and
the running of the `populate_published_groups` rake task, saving a
topic will publish an empty hash for the groups.

We will solve this condition in another way.

The TagPresenter spec had to be refactored to make the non-default
version work correctly. It now tests #render_for_publishing_api and
#build_groups separately test tags both with and without
published_groups.