Move topics to /topics namespace
Currently, topic pages appear at top level urls. This risks conflicts
with other pages, and pollutes the global namespace of the site with a
large number of pages.

This commit sets up everything to move to a /topic namespace.

- Set up redirects for old topic pages. The redirects are populated in
a database table and can be reused to allow renaming later.
- If the topic has any redirects, we show them on the topic page. This
helps the user understand why certain URLs work.
- Adds a ContentStoreHelpers module to the spec helpers. It implements
a fake ContentStore, that allows matchers to easily verify that the
correct data has been sent to the content store.