Send base_path instead of api_url to content-store
We want to update the frontend so that it can request tagged content
from rummager instead of contentapi. Rummager doesn't contain the
api_urls, so we'd have to fudge them from the base_path. This seems
wrong, and it would be better to send the base_paths in the curated
data. then we can match up the items by base_paths (both rummager and
contentapi have the base_path of the item).