Migrate to normalised table for redirects
A "redirect" can have multiple redirect routes. For example, the
"redirect" for /animal-welfare/pets has the routes:

/animal-welfare/pets -> /topic/animal-welfare/pets
/animal-welfare/pets/email-signup ->
/topic/animal-welfare/pets/email-signup

Currently the redirects holds both the redirect item as well as the
redirect routes. Because we want to add content_id to the redirects (but
not to the routes), we need to normalise the table.

We do the migration by using a different table_name for the old redirect
table and copying the data. In a later PR, the redirects table will be
removed and newest_redirects will be renamed to
redirects.