Add content_id to redirects
Redirects need their own `content_id` so that they can live in the
publishing-api.

We also add `replaced_content_item` to the item. This was added in
alphagov/govuk-content-schemas#142. It allows publishing-api to add
special behaviour for putting back redirects.

The tag_id is now mandatory, because we need a tag_id to send it to the
publishing-api. There is no UI to do this without specifying a tag, so
adding a database constraint is enough.