Enable disabled cops
These cops are disabled by default in govuk-lint, but without
specifying a reason. We enable them for this project, so that we can
see them in action in this new project. Whenever we encounter a cop
that doesn't work, we can disable it here and comment in a reason. We
can then port this back to govuk-lint.