Add download and parsing scripts for the alpha taxonomy
The alpha taxonomy is a work-in-progress taxonomical structure currently
being put together by GOV.UK information architects. In order to make
the taxonomy data available for prototyping of the new tagging interface,
a data pipeline is required to get taxons into the content store and link
to them from specific piece of content.

These additions handle the downloading and parsing of the remotely-stored
taxonomy mappings. These will be followed by additional tasks that will
handle interactions with the Publishing API and Content Store.