Solve content-items having invalid redirects
Collections-publisher currently sends invalid content-items to the
publishing-api. This is because some tags have redirect-routes that
don't "belong" to the tag. Eg, they have a path that doesn't start with
the tag's `base_path`:

For example:

https://collections-publisher.preview.alphagov.co.uk/topics/4505d908-89f
2-4322-956b-29ac243c608b

Has a redirect for `/benefits-credits`

This is illegal because `/benefits-credits` is not under the tag's base
path, `/topic/benefits-credits`.

`/topic/benefits-credits/yolo` would still be legal.

This PR adds a table that we can populate with these redirects. When
running the rake task, they will be sent indepdently, as full redirect
content-items to the publishing-api.

Note:
- The functionality introduced in
https://github.com/alphagov/collections-publisher/pull/163 is still
valid, and works like it should. It will generate redirect routes for
*itself*, so it will work with the publishing-api.
- `RedirectItem#related_tag` has no other purpose than making the data
migration reversible. It will be removed in the future.