Set the publishing app for taxons to 'collections-publisher'
Although the bulk import of taxons (and the mapping of content items to
taxons) is intended to sit within this application, editing of taxons
currently sits within collections-publisher. For convenience and speed, this
split in functionality will remain as we iterate on taxonomy managemement and
determine how best to structure things. In order for editing of taxons
to work correctly, we need to ensure the publishing app is set
appropriately to avoid validation errors from the publishing API.