Rewrite core feature specs
The feature specs describe the functionality of this application. Over
the last year this app has gotten more responsibilities, growing the
these tests.

It has some problems:

- Some are really large (~ 350 lines). It's hard to quickly scan which
functionality is being tested.
- They are hard to follow. Some of the tests have Cucumber-like
comments (When I make a change to the mainstream browse page), but
other don't.
- Lots of duplication in boilerplate code (create a record, navigate to
this page)
- Duplicated tests. Many specs test a slightly different code branch,
or test the same functionality twice.

hese problems make the test suite hard to work with. A small change can
break many tests at once, and it is hard to figure out why they are
failing. It is then challenging to figure out if the tests or the code
should change, resulting in shaky tests.

This PR rewrites the core feature specs for this app. It splits up the
functionality into features and scenarios, using methods to implement
the different steps. By combining examples (use the same scenario for
creating and editing a record) we reduce of code necessary.

This is a technique also used in content-tagger and explained in this
blog post:

https://about.futurelearn.com/blog/how-we-write-readable-feature-tests-w
ith-rspec/