Add wait_for_ajax helpers to prevent intermittently breaking topic curation feature
There's a race condition in the curation JS. When an item is dragged to
a list, it is 'reindexed' - if the item is newly placed, a ListItem is
created. If the item is reordered, the ListItem is updated.

When two drag and drops happen in quick succession (like they do in the
specs), a given item can be 'created' twice resulting in duplicate
entries in the list.

Until we solve this definitively in the client, this change is a band aid to
ensure that, when executing a drag and drop, the spec waits for the ajax
request to complete before continuing.