Upgrade Sidekiq to 3.X
This commit upgrades Sidekiq to 3.5.4, the latest version in the 3.X
series (4.0 has come out a few months ago). It is necessary to upgrade
because there are a couple of vulnerabilities reported in 2.X.

The issues  concerns these issues:

- https://github.com/mperham/sidekiq/pull/2422
- https://github.com/mperham/sidekiq/issues/2330
- https://github.com/mperham/sidekiq/pull/2309

As mentioned in the upgrade guide[1], we've upgradedA the Airbrake gem
as well. Other items in the upgrade guide are not applicable.

[1] https://github.com/mperham/sidekiq/blob/master/3.0-Upgrade.md