Allow tagging of collections-publisher documents
`collections-publisher` publishes topics and mainstream browse pages.
This commit allows those pages to be tagged.

This allows users, for example, to add link topics to other topics, and
add organisations to topics. There are no plans to do this, but we want
as little special cases as possible. It would be therefore be good to
treat the pages managed by collections-publisher as just a normal piece
of content on GOV.UK.

Trello: https://trello.com/c/GRt7NSbL