Use "minor" update type for tag publishing
Setting the update type to minor means we won't send email alerts for
changes to a topic title or description. This will become relevant if
topics will be tagged to other topics (we have no plans to do this, but
we need to be careful).

Since there is only ever one update type, all methods have been
replaced by a constant.

The only change is in the `RedirectItemPresenter` which previously was
sent a "republish" update type. This was incorrect, as the app is
publishing new data, not republishing old data.