Add support for blacklisting fields, and blacklist parent for TAP content
This adds functionality to switch off certain kinds of links in the
content tagger form, and sets this up specifically for 'parent' fields
when the document being tagged is owned by Travel Advice Publisher.

We need to be able to do this in order to migrate Travel Advice Publisher
to the new tagging architecture. Currently, the parent link for TAP
content is set by default to be an array of hashes modelling a
breadcrumb. This isn't something we're able to handle in content-tagger
right now. The workaround in this commit allows us to migrate TAP away
from Panopticon while we determine how best to handle this on an ongoing
basis.