Use lookup endpoint to find the content_id
We recently introduced a `/lookup-by-base-path` endpoint that can be
used to find the content_id for any base_path. This PR uses that
endpoint to do the lookup, which will enable us to no longer depend on
the content-store for this application.

It does remove the feature where we tell the user that a content-item
can't be tagged because it's a redirect or gone item. We will replace
this by a warning in the tagging-screen.