Add integration test for tagging to restricted pages
This adds an integration tests to test the scenario where the user tags
a page that doesn't allow tagging to all link types. For example,
travel-advice-publisher, which doesn't allow `parent` tagging.

I've chosen to do this in a separate file to keep the core feature
tests clean. This does introduce a number of duplicated steps, but I
feel that duplication & explicitness is preferable here to extracting
steps to a common module.