Fix item titles in curated lists
Item titles for display in collections-publisher are currently looked up from data
persisted to the mysql db at list curation time.
This is not consistent with the collections frontend, which looks up titles from rummager.

It is possible for an item to appear in a curated list but to since have had its tag removed
or been deleted. In this case the collections frontend doesn't display the item.

Also, it is possible for a content item to have been renamed since it was included in a
curated list. It can then be added again under its new title.
In this case, collections-publisher will display two items which, while pointing to the
same base_path, appear different becase the persisted titles are different.
Collections frontend will display both items using the most recent title from rummager,
giving the appearance of duplicates which cannot be found in collections-publisher.

See https://trello.com/c/gfzHlBUn/620-collections-publisher-duplicates-in-list-editor-issue

Finally, it is possible for the persisted title to be null, in which case
collections-publisher can display an empty <li> element.

Once all taggable content is being published via publishing-api, we can dispense with
the rummager lookup in favour of publishing-api.

In the meantime, this commit fixes the display issues in collections-publisher by using
the rummager title where it exists, and falling back to the persisted title or placeholder
text where it does not.