Rename TaxonForm.parents to TaxonForm.taxon_parents
The schema for Taxon will have a new link type called `taxon_parents`. This
change is so we are clear about what parents we are referring to when calling
the `taxon_parents` method.

Furthermore, introducing `instance_double` so we can verify method calls to the
doubles and avoid having method calls no longer present in the actual
implementation. This makes sure we call `taxon_parents` in TaxonFetcher.