Refer to the newly created parent_taxons from the API
As Taxons have a new link type called parent_taxons that holds information
about its multiple parents, we should from now on use it to create/update
taxons. This way, we will be able to have multiple parents and stop
interfering with the existing link type parent, which is used for breadcrumbs
and does not allow more than one item in the list.