Use publishing-api endpoint for linked items
Publisher apps should never use the content-store, because it's not
synchronous and could cause confusion for users. In this case, taggings
would not immediately show up because publishing-api sends the tags to
the content-store via a Sidekiq worker.

This commit replaces the call to content-store with a call to the
publishing-api so that we no longer depend on the former.