Use external class to build TaxonForm objects
This is an attempt at simplifying the TaxonForm class, which has many
responsibilities.

There was sufficient logic in the `build` class method to justify its own class.

The builder has to fetch both the content and its links from the publishing API,
and then perform logic on assigning parent taxons as well.

The resulting object is an initialized TaxonForm from existing data.