Make initial import action asynchronous
At the moment, when the user creates a new spreadsheet import by
submitting a URL, the application will start the initial import step
(parsing the CSV, creating records for each row) in the request cycle.

This is slow and does not allow us to add more expensive validation to
this step.

This commit is the minimum we have to do to move the initial import
action to a Sidekiq worker.

- It adds a very crude state thing, which could perhaps be replaced by
a real state machine.
- The workflow is a bit clunky now. For example, it's not clear to the
user what state the spreadsheet is in. We'll address this in follow up
PRs (https://trello.com/c/7tcrUgcz).

Trello: https://trello.com/c/xi83pGTZ