Make sure we mark TaggingSpreadsheets as deleted
Instead of deleting TaggingSpreadsheets, we should mark them as deleted and keep
a record of them in the backend for audit purposes. As they perform actions on
content items, we should probably keep a record of them in case something goes
wrong.

This change makes sure we don't perform a hard-delete on those spreadsheets. We
instead mark them as deleted and remove those ones from the index page.