Fix tests after re-arrangement of redirects for tagging spreadsheets
Commit number 833f2a changed the redirects of some controller actions in order
to improve the user experience. Unfortunately, jenkins didn't run the tests for
some reason and I ended up merging the PR. Some of the tests were broken because
they relied on a flow that doesn't exist anymore, so this commit fixes it.