Add visual differences to new and edit taxon pages
It is not clear if we are creating or editing a taxon from a user perspective.
This commit makes sure we have callout divs on the page with color and text
indications, making sure we differentiate between the 2 pages.

The edit page also has a confirmation box to force the user to confirm the
changes to the taxon.

Trello: https://trello.com/c/CMul4Ru3/112-improve-how-we-differentiate-between-the-new-taxon-and-edit-taxon-forms-s