Change submit buttons to have class "success"
According to the style guide:
When submitting a form that will create or update items use btn-success.

For everything else:
Buttons should always be the default grey unless there is a good reason
otherwise. Buttons that initiate a user action (ie lead to a form, pop-up a
modal) should be grey.

This commit makes sure that all submit buttons have the correct CSS class.