Allow exporting taxons
We would like to allow users to export taxons in order to make it easier to
create tagging spreadsheets.

Those spreadsheets need a mapping between taxon names and their content IDs. The
CSV file we are generating gives us exactly that, together with the link type.
This way, we can download a list of user-selected taxons into a format expected
by the tagging spreadsheet.

More information in trello: https://trello.com/c/gRoOrJoo/121-spike-build-functionality-into-content-tagger-which-makes-creating-a-tag-importer-spreadsheet-easier