Allow users to copy/paste taxons from HTML view
We would like to allow users to export taxons in order to make it easier to
create tagging spreadsheets.

Those spreadsheets need a mapping between taxon names and their content IDs.
This commit adds a link from the taxons index page into a new view which
contains a table with the necessary columns: title, content ID and link type.

The user can then copy that HTML page into a spreadsheet and it would work as if
we copied CSV data.