User Taxon objects as opposed to Hashes
We already have a model to represent a Taxon. Currently, we sometimes use the
model (e.g. from the TaxonBuilder service class), but in other situations (when
fetching taxons via the TaxonFetcher service class) we use Hashes.

This commit makes sure the method `taxons` in TaxonFetcher returns instances of
Taxons instead of hashes.