Convert TagMapping into a polymorphic relation
We now allow a TagMapping to belong to anything that creates a `tagging_source`
relation with it. This way we can have different objects having associations
with TagMapping.

This is particularly useful in the case where `TaggingSpreadsheet` and a future
`BulkTagging` will have `has_many` relationships with `TagMapping`.

`TagMapping` will represent a specific content item being tagged with a specific
link type, which is generic enough for the multiple tagging routes one could
take.