Order taxons by public_updated_at
This commit sorts taxons by their `public_updated_at` attribute rather than by
`title`.

This is now possible because we are using `get_content_items` instead of
`get_linkables` to fetch taxons, which returns a lot more attributes for each of
the taxons.

We are sorting records via the Publishing API in order to allow us to use
pagination in the future without breaking this functionality.