Don't mutate tagging_spreadsheet injected into FetchRemoteData
Prior to this commit, the tagging_spreadsheet was used to build its own
tag_mappings via the association. This was problematic - if the
tag_mappings being created were invalid for whatever reason, any attempt
at updating the tagging spreadsheet outside of this class (putting it
into an errored state, for example) would result in failure as it
attempts to persist the tag_mappings again.

This change ensures there isn't any mutation of the injected tagging
spreadsheet.