Workaround for linkables with internal names of nil
Accessing linkables on the Publishing API can return nil for fields such
as internal_name or title when the linkable item is a redirect. In this
instance we have a problem when the internal_name is nil as we require
it to be a string.

To address this we filter the linkables from the Publishing API to only
include instances where the internal_name is not nil.