Fix migration for polymorphic tag mappings
The `taggable_type` attribute got renamed from `taggable_type`.

Because locally there weren't any records in the tag_mappings table,
this didn't surface. On staging there are records, so this migration
fails because `taggable_type` doesn't exist.

Also makes it use a SQL query, which is faster and more robust.