Add feature spec to test the progress bar
This test makes sure we can see the progress bar and that it completes.

The side effects of adding this feature test (and other JS tests in the future)
is that:
- we cannot use `:rack_test`, therefore we need to rely on a slower `webkit`
  driver to run those particular javascript tests;
- we cannot enable RSpec fixtures, and have to use `database_cleaner` instead;
- we have to use a `truncation` strategy, which means we then need to create our
  User record before each example (or re-create it after the js test)

On the plus side, we can select the Capybara javascript driver depending on the
type of tests, which means most of the tests still use `:rack_test`.