Display tagging spreadsheet errors under URL
Previously we displayed the errors in the description column. Because the URL
column will take most of the table space, the error message was squashed in a
smaller column, making it harder to read.

This commit moves the error messages to the URL column, where there is more
space to show it.