Refactor how we update links via PublishLinksWorker
We delegated the responsibility of updating the tag mappings with success or
error to the PublishLinks service object. Although this works well, it doesn't
let us reuse PublishLinks in other contexts to update links of content items.

This change aims at moving the marking as success/error logic into the
PublishLinksWorker instead, where it is actually needed. Furthermore, this
change makes sure PublishLinks is a generic service object to update and publish
links on any content item.