Allow PublishLinks to delete link types
When specifically giving an empty array for a given link type, PublishLinks will
delete the existing links of that link type.

As an example, if we call the PublishLinks service with an object which includes
a `links_to_update` method like this:

```
def links_to_update
  { parent: [] }
end
```

We will then override the `parent` link with an empty list.