Add pagination to index page of Taxons
This method adds pagination links to the Taxon index page via the `kaminari`
gem.

In order to allow this, a new `search` method is available on `RemoteTaxons`
that allows us to search for all taxons with pagination information.

The results are then wrapped in a model class with the interface that `kaminari`
understands.