Use linkables for tag migrations dropdown
This changes the way we populate the select-a-taxon dropdown in the tag
migration screen. Linkables is the appropriate class / API call for
this, since it's already used for other dropdowns. This means that it
uses the `internal_name` property (which disambiguates taxons with the
same name) and shows the taxons alphabetically.

Note that `build_linkable` only returns random content by default. I
think that's better than hard coded values, since it requires the tests
to be explicit about what they expect the helper method to produce.
Perhaps basic_content_item should be rewritten to to something like it.