Change workflow based on product and design feedback
This commit changes the flow of some of the existing features in content-tagger.
See the Wiki for more details on these changes.

It also makes use of I18n for link titles in order to minimise test failures on
copy changes.

Kudos to @tijmenb and @mobaig that have done most of the work.