Add feature to move content items between taxons
This feature allows us to move content items between taxons. It is
similar to bulk tagging, however it deletes the link of the source
taxon, keeping only the link to the destination taxon.

Example:

If we have tagged both `Content Item 1` and `Content Item 2` with `Taxon
A`, when we move the content items over to `Taxon B` using this feature,
the original link to `Taxon A` will be deleted. Both content items will
only keep the link to `Taxon B`.

Both taxons are left untouched, i.e. none is deleted.

Trello:
https://trello.com/c/YRnivAny/214-content-tagger-allow-ias-to-move-content-from-taxon-to-taxon