Perform atomic link updates
Before this change, we were grouping tag mappings into an intermediate data
structure in order to group link updates by base path before sending it to the
publishing API.

Although this has been working well, it's not very flexible and makes it hard to
change and reason about.

This commit removes that logic so that we perform atomic link updates against
the Publishing API. The consequence of this change is that it is now clearer
what the system does and at the same time it allows us to plug in other pieces
of logic such as removing individual links (from the "Move Taxons" feature, for
example). Furthermore, each tag mapping is executed on its own, meaning it's
easier to debug what the errors are.

The downside of this change is that we will be doing a few more API calls to the
Publishing API. However, I don't anticipate this aggregation to save us a lot of
requests, since we won't be tagging content items with many taxons at the same
time.

Something else to be aware of - if there are repeated link updated in a tagging
spreadsheet, this approach will execute all of those. This issue should be
minimal and won't affect anything whatsoever apart from a surplus request.

This change is possible because:
- we already append link updates to existing links instead of overriding them;
- we already aggregate the link updates in the view so people know what changes
  there are.