Taxon#notes_for_editors needs to be a string
When sending a taxon to publishing api, it needs to have the attribute `notes_for_editors` as a string.
At the moment it was sending a `null` value.

This commit aims to fix it to comply with the [schema definition](https://github.com/alphagov/govuk-content-schemas/blob/master/formats/taxon/publisher/details.json#L11)