Use internal_name from details hash
The publishing API used to return the internal_name field both in the details
hash and in the top level content item hash. The latter has been removed from
the publishing API and we should not use it.

This commit looks up internal names in the details hash and fixes the tests so
we always look in the details hash.