Move "Bulk tag by upload" within "Bulk Tag"
We belive this functionality should move into the Bulk Tag tab.

Although having it directly within the page would make it confusing when
you search for content to be tagged and also see the list that comes
from the uploaded content. That's why I've created a "button" to point
to the "Bulk tag by upload"

Also the reason to make it a button is that it stands out in the eye of
the user instead of a link, which looks tiny and reads weird next to
"Bulk tag history".