Update taxonomy visualisation
This commit updates the front end code and underlying data structure of
the taxonomy visualisation feature. Previously we could only see a basic
tree structure detailing the child taxons and immediate parents of a
given taxon. The updated design allows us to see all ancestors and all
successors in a clearer layout.

- Update ExpandedTaxonomy so that it now provides an abstraction over
  two trees - the fully expanded set of parents and the fully expanded
  set of children. Also includes convenience methods for querying data
  about the root node and its immediate parents/children.
- Rework the HTML and CSS to accommodate the new design and additional
  data being displayed.
- Update the viewing_taxons feature spec to reflect the new behaviour.
- Some minor naming clarification in the TreeNode class.