Refactor ExpandedTaxonomy to provide more granular API
Make it possible to expand either the child nodes or parent nodes in
isolation. Client code may require just one of these and there's a
performance advantage in being able to call them independently. This is
particularly the case when you only need parent nodes - a larger number
of network requests are required to build out the child node expansion.